Expression: Add new test sub-module for finders and expr_visitors
